Refactor risky-octal and improve its messaging
Fixes: #2758
